apps/api/requirements.txt
# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20

# Data Validation
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.30.0

# HTTP Clients
httpx>=0.28.0
aiohttp>=3.10.0

# Auth & Security
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
bcrypt>=4.0.0

# Media Processing
Pillow>=10.0.0
pydub>=0.25.1
numpy>=1.26.0

# Google Cloud (for Gemini)
google-generativeai>=0.5.0
google-cloud-storage>=2.14.0
